+++ src/all.ts
@@ -2,6 +2,7 @@
 export { CpsBadge } from './components/badge';
 export { CpsButton } from './components/button';
 export { CpsButtonGroup } from './components/button-group';
+export { CpsCheckbox } from './components/checkbox';
 export { CpsChip } from './components/chip';
 export { CpsIcon } from './components/icon';
 export { CpsIconButton } from './components/icon-button';