+++ CHANGELOG.md
@@ -8,6 +8,7 @@ O versionamento deste projeto é aderente aos princípios de [Semantic Versionin
 
 ### Added
 
+- Criação do componente `<cps-checkbox>`.
 - Criação do componente `<cps-textarea>`.
 - Documentação detalhando os fundamentos de utilização de elementos customizados.